Add files via upload
I am conducting an experiment to test whether adding second-degree connections can improve the robustness of a scale-free network. The approach is as follows: First, I created a scale-free network with 1000 nodes. Then, I added second-degree connections to the top 5% of nodes based on degree centrality. After that, I conducted an attack targeting the top 10% of nodes based on degree centrality to observe the robustness of the original network compared to the network with added second-degree connections. However, when I changed the 14th line of my code to copy the network, the robustness of both networks became identical, which is obviously not normal. I'm not sure where the problem in my code lies. Could you please help me? Thank you."